Dashing
Dashing is now added that can let Sizzle go over an edge if they are close enough but will stop them shy of going over if it is too close to the end of the dash. Moving Sizzle was easy but I added a system, gizmos shown a bit later in the video, that if the player just almost dashes over an edge, ie doesn't reach that green threshold, the system will correct the player and add some distance to the dash so the tail fits into the red cube. This means that we can set up dashing of edges to expect to fit the whole body so it doesn't end with the tail clipping in geometry, and it doesn't require the player to perfectly make the dash.

TLDR: Dashing movement and system to help make it easier over edges