Added Spark Shoot Event
I added in a basic "Play_Sparks" event that will occur in the demo scene. This currently is just a component based SFX, however the individual pops of the SFX could likely be tied to an event. Maybe we can test it with the current event and see how that goes?